+++ src/main/java/com/clinicAnimal/system_vet_clinica/Web/dto/request/CitaRequestDTO.java
@@ -1,4 +1,4 @@
-package com.clinicAnimal.system_vet_clinica.web.dto;
+package com.clinicAnimal.system_vet_clinica.web.dto.request;
 
 
 import lombok.*;
