+++ src/main/java/com/clinicAnimal/system_vet_clinica/Web/dto/request/PersonalRequestDTO.java
@@ -1,11 +1,9 @@
-package com.clinicAnimal.system_vet_clinica.web.dto;
+package com.clinicAnimal.system_vet_clinica.web.dto.request;
 
 import com.clinicAnimal.system_vet_clinica.domain.model.Estados.EstadoPersonal;
 import lombok.*;
 
-import java.time.LocalDate;
 import java.time.LocalTime;
-
 
 @Getter
 @Setter
@@ -13,18 +11,17 @@ import java.time.LocalTime;
 @NoArgsConstructor
 @Builder
 
-public class PersonalResponseDTO {
-    private Long perId;
+public class PersonalRequestDTO {
     private String nombrePersonal;
     private String apellidoPersonal;
     private int edad;
     private String dni;
     private String email;
     private String telefono;
-    private LocalDate fechaRegistro;
     private LocalTime fechaEntrada;
     private LocalTime fechaSalida;
     private EstadoPersonal estadoPersonal;
+    //private Long rolCodigo;
     private String rolNombre;
     private String nomArea;
 }
