+++ src/main/java/com/clinicAnimal/system_vet_clinica/Web/dto/response/CitaResponseDTO.java
@@ -1,4 +1,4 @@
-package com.clinicAnimal.system_vet_clinica.web.dto;
+package com.clinicAnimal.system_vet_clinica.web.dto.response;
 
 
 import com.clinicAnimal.system_vet_clinica.domain.model.Estados.EstadoCita;
