+++ src/main/java/com/clinicAnimal/system_vet_clinica/Web/dto/response/RolesResponseDTO.java
@@ -1,4 +1,4 @@
-package com.clinicAnimal.system_vet_clinica.web.dto;
+package com.clinicAnimal.system_vet_clinica.web.dto.response;
 import lombok.*;
