+++ src/main/java/com/clinicAnimal/system_vet_clinica/Web/dto/response/ServicioResponseDTO.java
@@ -1,0 +1,18 @@
+package com.clinicAnimal.system_vet_clinica.web.dto.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ServicioResponseDTO {
+    private Long id;
+    private String nombre;
+    private String descripcion;
+    private Integer duracion;
+    private double precio;
+    private Long idArea;
+    private String nomArea;
+}
