+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/mapper/AreasMap.java
@@ -1,9 +1,8 @@
 package com.clinicAnimal.system_vet_clinica.aplication.mapper;
 
 import com.clinicAnimal.system_vet_clinica.domain.model.Areas;
-import com.clinicAnimal.system_vet_clinica.domain.model.Personal;
-import com.clinicAnimal.system_vet_clinica.web.dto.AreasRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.AreasResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.AreasRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.AreasResponseDTO;
 
 public interface AreasMap {
     Areas toEntity(AreasRequestDTO dto);
