+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/mapper/CitaMap.java
@@ -2,8 +2,8 @@ package com.clinicAnimal.system_vet_clinica.aplication.mapper;
 
 import com.clinicAnimal.system_vet_clinica.domain.model.Cita;
 import com.clinicAnimal.system_vet_clinica.domain.model.Mascota;
-import com.clinicAnimal.system_vet_clinica.web.dto.CitaRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.CitaResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.CitaRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.CitaResponseDTO;
 
 public interface CitaMap {
     Cita toEntity(CitaRequestDTO dto, Mascota mascota);
