+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/mapper/PersonalMap.java
@@ -3,8 +3,8 @@ package com.clinicAnimal.system_vet_clinica.aplication.mapper;
 import com.clinicAnimal.system_vet_clinica.domain.model.Areas;
 import com.clinicAnimal.system_vet_clinica.domain.model.Personal;
 import com.clinicAnimal.system_vet_clinica.domain.model.Roles;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.PersonalRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.PersonalResponseDTO;
 
 public interface PersonalMap {
     Personal toEntity(PersonalRequestDTO dto, Roles roles, Areas areas);
