+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/mapper/RecetaMap.java
@@ -2,8 +2,8 @@ package com.clinicAnimal.system_vet_clinica.aplication.mapper;
 
 import com.clinicAnimal.system_vet_clinica.domain.model.Receta;
 import com.clinicAnimal.system_vet_clinica.domain.model.Cita;
-import com.clinicAnimal.system_vet_clinica.web.dto.RecetaRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.RecetaResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.RecetaRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.RecetaResponseDTO;
 import com.clinicAnimal.system_vet_clinica.web.dto.RecetaUpdateDTO;
 
 public interface RecetaMap {
