+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/mapper/impl/CitaMapImpl.java
@@ -4,8 +4,8 @@ import com.clinicAnimal.system_vet_clinica.aplication.mapper.CitaMap;
 import com.clinicAnimal.system_vet_clinica.domain.model.Cita;
 import com.clinicAnimal.system_vet_clinica.domain.model.Estados.EstadoCita;
 import com.clinicAnimal.system_vet_clinica.domain.model.Mascota;
-import com.clinicAnimal.system_vet_clinica.web.dto.CitaRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.CitaResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.CitaRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.CitaResponseDTO;
 import org.springframework.stereotype.Component;
 
 @Component
