+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/mapper/impl/RolesMapImpl.java
@@ -2,8 +2,8 @@ package com.clinicAnimal.system_vet_clinica.aplication.mapper.impl;
 
 import com.clinicAnimal.system_vet_clinica.aplication.mapper.RolesMap;
 import com.clinicAnimal.system_vet_clinica.domain.model.Roles;
-import com.clinicAnimal.system_vet_clinica.web.dto.RolesRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.RolesResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.RolesRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.RolesResponseDTO;
 import lombok.Builder;
 import org.springframework.stereotype.Component;
