+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/service/CitaService.java
@@ -1,8 +1,8 @@
 package com.clinicAnimal.system_vet_clinica.aplication.service;
 
 import com.clinicAnimal.system_vet_clinica.domain.model.Estados.EstadoCita;
-import com.clinicAnimal.system_vet_clinica.web.dto.CitaRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.CitaResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.CitaRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.CitaResponseDTO;
 import com.clinicAnimal.system_vet_clinica.web.dto.CitaUpdateDTO;
 
 import java.util.List;
