+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/service/Impl/AreasServImpl.java
@@ -5,8 +5,8 @@ import com.clinicAnimal.system_vet_clinica.aplication.mapper.AreasMap;
 import com.clinicAnimal.system_vet_clinica.aplication.service.AreaService;
 import com.clinicAnimal.system_vet_clinica.domain.model.Areas;
 import com.clinicAnimal.system_vet_clinica.domain.repository.AreasRep;
-import com.clinicAnimal.system_vet_clinica.web.dto.AreasRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.AreasResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.AreasRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.AreasResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
