+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/service/Impl/CitaServImpl.java
@@ -6,8 +6,8 @@ import com.clinicAnimal.system_vet_clinica.domain.model.Estados.EstadoCita;
 import com.clinicAnimal.system_vet_clinica.domain.model.Mascota;
 import com.clinicAnimal.system_vet_clinica.domain.repository.CitaRepository;
 import com.clinicAnimal.system_vet_clinica.domain.repository.MascotaRepository;
-import com.clinicAnimal.system_vet_clinica.web.dto.CitaRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.CitaResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.CitaRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.CitaResponseDTO;
 import com.clinicAnimal.system_vet_clinica.web.dto.CitaUpdateDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
