+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/service/Impl/PersonalServImpl.java
@@ -9,8 +9,8 @@ import com.clinicAnimal.system_vet_clinica.domain.model.Roles;
 import com.clinicAnimal.system_vet_clinica.domain.repository.AreasRep;
 import com.clinicAnimal.system_vet_clinica.domain.repository.PersonalRep;
 import com.clinicAnimal.system_vet_clinica.domain.repository.RolesRep;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.PersonalRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.PersonalResponseDTO;
 import com.clinicAnimal.system_vet_clinica.web.dto.PersonalUpdateDTO;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
