+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/service/Impl/RecetaServImpl.java
@@ -5,8 +5,8 @@ import com.clinicAnimal.system_vet_clinica.domain.model.Cita;
 import com.clinicAnimal.system_vet_clinica.domain.model.Receta;
 import com.clinicAnimal.system_vet_clinica.domain.repository.CitaRepository;
 import com.clinicAnimal.system_vet_clinica.domain.repository.RecetaRepository;
-import com.clinicAnimal.system_vet_clinica.web.dto.RecetaRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.RecetaResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.RecetaRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.RecetaResponseDTO;
 import com.clinicAnimal.system_vet_clinica.web.dto.RecetaUpdateDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
