+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/service/Impl/RolesServImpl.java
@@ -3,8 +3,8 @@ package com.clinicAnimal.system_vet_clinica.aplication.service.Impl;
 import com.clinicAnimal.system_vet_clinica.aplication.mapper.RolesMap;
 import com.clinicAnimal.system_vet_clinica.domain.model.Roles;
 import com.clinicAnimal.system_vet_clinica.domain.repository.RolesRep;
-import com.clinicAnimal.system_vet_clinica.web.dto.RolesRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.RolesResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.RolesRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.RolesResponseDTO;
 import com.clinicAnimal.system_vet_clinica.aplication.service.RolesService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
