+++ src/main/java/com/clinicAnimal/system_vet_clinica/aplication/service/PersonalService.java
@@ -1,13 +1,9 @@
 package com.clinicAnimal.system_vet_clinica.aplication.service;
 
 
-import com.clinicAnimal.system_vet_clinica.domain.model.Roles;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalResponseDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.RolesResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.PersonalRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.PersonalResponseDTO;
 
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalResponseDTO;
 import com.clinicAnimal.system_vet_clinica.web.dto.PersonalUpdateDTO;
