+++ src/main/java/com/clinicAnimal/system_vet_clinica/domain/model/Cliente.java
@@ -30,7 +30,6 @@ public class Cliente {
     private String direccion;
     @Column(name="email")
     private String email;
-
     @OneToMany(mappedBy = "cliente")
     private List<Mascota> mascotas;
