+++ src/main/java/com/clinicAnimal/system_vet_clinica/web/controller/AreaController.java
@@ -2,8 +2,8 @@ package com.clinicAnimal.system_vet_clinica.web.controller;
 
 
 import com.clinicAnimal.system_vet_clinica.aplication.service.AreaService;
-import com.clinicAnimal.system_vet_clinica.web.dto.AreasRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.AreasResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.AreasRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.AreasResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
