+++ src/main/java/com/clinicAnimal/system_vet_clinica/web/controller/PersonalController.java
@@ -1,8 +1,8 @@
 package com.clinicAnimal.system_vet_clinica.web.controller;
 
 import com.clinicAnimal.system_vet_clinica.aplication.service.PersonalService;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalRequestDTO;
-import com.clinicAnimal.system_vet_clinica.web.dto.PersonalResponseDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.request.PersonalRequestDTO;
+import com.clinicAnimal.system_vet_clinica.web.dto.response.PersonalResponseDTO;
 import com.clinicAnimal.system_vet_clinica.web.dto.PersonalUpdateDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
